date formatting and profile name bug fix

+++ app/src/main/java/uk/ac/tees/w9543466/pathlight/worker/WorkDto.java
@@ -58,7 +58,7 @@ public class WorkDto {
         return startTime;
     }
 
-    public void setStartTime(double startTime) {
+    public void setStartTime(long startTime) {
         this.startTime.set(startTime + "");
     }
